Write unit test:

+++ src/graphql/resolvers/index.js
@@ -39,20 +39,15 @@ export const resolvers = {
       const male = male_population && parseInt(male_population, 10)
       const total = female_population && male_population && female + male;
       const id = randomNumber(10);
+      if (!location) {
+        throw new Error('Failed to create location because location field is required')
+      }
       const newlyCreated = new LocationModel({ id, location, male_population: male, female_population: female, total_population: total, parent_location })
-      const findLocation = await LocationModel.findOne( { id: { $eq: parseInt(parent_location, 10) } } )
-      if(!findLocation){
+      const findLocation = parent_location ? await LocationModel.findOne({ id: { $eq: parseInt(parent_location, 10) } }) : { location: "exist" }
+      if (!findLocation) {
         throw new Error('The parent location does not exist')
       }
-      if (!location) {
-        validationErrors.location = "The location field can not be null"
-      }
-      if (Object.keys(validationErrors).length > 0) {
-        throw new UserInputError(
-          'Failed to create locations because there was an error in the location input',
-          { validationErrors }
-        );
-      }
+
       const newLocation = await newlyCreated.save()
       return newLocation;
     },
@@ -62,13 +57,21 @@ export const resolvers = {
       if (!id) {
         throw new Error('Can not update because id is required')
       }
+      const findLocation = await LocationModel.findOne({id})
+      if(!findLocation) {
+        throw new Error('The location does not exist')
+      }
       const updatedLocation = await LocationModel.findOneAndUpdate({ id }, { $set: args }, { new: true });
       return updatedLocation
     },
 
-    deleteLocation: async(root, args) => {
+    deleteLocation: async (root, args) => {      
       if (!id) {
         throw new Error('Can not delete location because id is required')
+      }
+      const findLocation = await LocationModel.findOne(args)
+      if(!findLocation) {
+        throw new Error('Can not delete location because it does not exist')
       }
       const deleted = await LocationModel.findOneAndRemove(args)
       return deleted;
